Repair tests, add default blob type.
