+++ testSuite/scripts/run.py
@@ -13,6 +13,7 @@ from test_blob_sync import *
 from test_service_to_service_copy import *
 from test_blobfs_download_SAS import *
 from test_blobfs_upload_SAS import *
+from test_autodetect_blob_type import *
 import glob, os
 import configparser
 import platform
@@ -194,7 +195,8 @@ def main():
                          FileShare_Upload_User_Scenario,
                          BlobFs_Upload_ShareKey_User_Scenarios,
                          BlobFs_Download_SharedKey_User_Scenarios,
-                         Service_2_Service_Copy_User_Scenario]
+                         Service_2_Service_Copy_User_Scenario,
+                         Autodetect_Blob_Type_Scenario]
     suites_list = []
 
     loader = unittest.TestLoader()
